+++ zhalilov.rail/T1/polygon.cpp
@@ -1,4 +1,4 @@
-#include "Polygon.hpp"
+#include "polygon.hpp"
 
 #include <stdexcept>
 #include <algorithm>
