+++ zhalilov.rail/T1/polygon.hpp
