+++ zhalilov.rail/T1/rectangle.hpp
